backend/requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-cors-headers==3.14.0
django-import-export==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.1
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pycparser==2.21
PyJWT==2.6.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
urllib3==1.26.14
django-cors-headers==3.14.0
coverage==7.2.1
xlrd==2.0.1
xlwt==1.3.0
django-environ==0.10.0
psycopg2-binary==2.9.3